fix: add correct import after refactor
